Creating prefabs for vendors & audio adjustment

+++ Metroidvania_Game/Assets/ScriptableObjects/SceneInfo.cs
@@ -19,7 +19,7 @@ public class SceneInfo : ScriptableObject
 
     public bool isHPBought = false;
     public bool isAxeUsed = false;
-    public int swordDamageValue = 1;
+    public float swordDamageValue = 1;
     public float knockbackForce = 1;
 
     [Space(20)]
